requirements.txt
Flask==2.0.0
Flask-Cors==4.0.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
pydantic

pymediainfo
imageio
uiautomation>=2
requests
oss2
python-dotenv
colorlog
flask
gunicorn
pywin32